Sorted out API controller inheritance
