+++ app/controllers/concerns/cache_utils.rb
@@ -14,4 +14,9 @@ module CacheUtils
   def clear_cache_for_regexp(regexp)
     Cache::Cleaner.clear_cache_for_regexp(regexp)
   end
+
+  def set_caching_headers
+    return true if Rails.env.development?
+    expires_in 15.minutes, private: true, 's-maxage' => 2.hours
+  end
 end
