+++ app/controllers/concerns/exceptions.rb
@@ -1,0 +1,9 @@
+module Exceptions
+  extend ActiveSupport::Concern
+
+  included do
+    rescue_from ActionController::ParameterMissing do |exception|
+      render json: {error: exception.message}, status: 400
+    end
+  end
+end
